move `NetworkRelaxationScenarioGenerator` to scripts
